refactor(StationController): 컨트롤러에서 repository를 필드로 들고있는 것 수정
